Rename electronBuilder configs
