+++ desktop/README.md
@@ -87,7 +87,7 @@ mc policy set public electron-builder/electron-builder
 Once you have Min.IO setup and running, the next step is to temporarily revert some changes from https://github.com/Expensify/App/commit/b640b3010fd7a40783d1c04faf4489836e98038d, specifically
 
 1. Update the `desktop-build` command in package.json to add `--publish always` at the end
-2. Update electronBuilder.config.js to replace the `publish` value with the following:
+2. Update electronBuilder.ghactions.config.js to replace the `publish` value with the following:
 ```
  publish: [{
    provider: 's3',
