+++ package.json
@@ -18,10 +18,10 @@
     "web": "node web/proxy.js & webpack-dev-server --open --config config/webpack/webpack.dev.js",
     "build": "webpack --config config/webpack/webpack.prod.js",
     "build-staging": "webpack --config config/webpack/webpack.staging.js",
-    "desktop-build": "export NODE_ENV=production && webpack --config config/webpack/webpack.prod.js --platform desktop && electron-builder --config config/electronBuilder/electronBuilder.config.js",
-    "desktop-build-staging": "export NODE_ENV=staging && webpack --config config/webpack/webpack.staging.js --platform desktop && electron-builder --config config/electronBuilder/electronBuilder.config.js",
-    "desktop-build-local": "export NODE_ENV=production && webpack --config config/webpack/webpack.prod.js --platform desktop && electron-builder --config config/electronBuilder/electronBuilder.base.config.js",
-    "desktop-build-staging-local": "export NODE_ENV=staging && webpack --config config/webpack/webpack.staging.js --platform desktop && electron-builder --config config/electronBuilder/electronBuilder.base.config.js",
+    "desktop-build": "export NODE_ENV=production && webpack --config config/webpack/webpack.prod.js --platform desktop && electron-builder --config config/electronBuilder/electronBuilder.ghactions.config.js",
+    "desktop-build-staging": "export NODE_ENV=staging && webpack --config config/webpack/webpack.staging.js --platform desktop && electron-builder --config config/electronBuilder/electronBuilder.ghactions.config.js",
+    "desktop-build-local": "export NODE_ENV=production && webpack --config config/webpack/webpack.prod.js --platform desktop && electron-builder --config config/electronBuilder/electronBuilder.local.config.js",
+    "desktop-build-staging-local": "export NODE_ENV=staging && webpack --config config/webpack/webpack.staging.js --platform desktop && electron-builder --config config/electronBuilder/electronBuilder.local.config.js",
     "ios-build": "fastlane ios build",
     "android-build": "fastlane android build",
     "test": "jest",
